Stack: move functions for manipulating heads to own module
