+++ crates/gitbutler-stack/src/lib.rs
@@ -1,3 +1,4 @@
+mod heads;
 mod series;
 mod stack;
 pub use series::Series;
